chore(storybook): migrated storybook to seperate directory
