+++ apps/storybook/svelte.config.js
@@ -1,0 +1,5 @@
+import { vitePreprocess } from '@astrojs/svelte'
+
+export default {
+  preprocess: vitePreprocess(),
+}
