--- apps/website/.netlify/v1/config.json
@@ -1,1 +1,0 @@
-{"images":{"remote_images":[]},"headers":[{"for":"/_astro/*","values":{"Cache-Control":"public, max-age=31536000, immutable"}}]}
