+++ apps/website/src/components/ui/button/index.ts
@@ -1,5 +1,5 @@
-import { type VariantProps, tv } from 'tailwind-variants'
 import type { Button as ButtonPrimitive } from 'bits-ui'
+import { type VariantProps, tv } from 'tailwind-variants'
 import Root from './button.svelte'
 
 const buttonVariants = tv({
@@ -37,12 +37,12 @@ type Props = ButtonPrimitive.Props & {
 type Events = ButtonPrimitive.Events
 
 export {
-  Root,
-  type Props,
-  type Events,
   //
   Root as Button,
-  type Props as ButtonProps,
-  type Events as ButtonEvents,
+  Root,
   buttonVariants,
+  type Events as ButtonEvents,
+  type Props as ButtonProps,
+  type Events,
+  type Props,
 }
